Devices
Added a Manufacturer, DeviceModel, and a Device.
Which should cover all necessary fields needed for a flexible device.

Switched the relationship between a device and a user to a polymorphic
to add the ability to be assigned to multiple things at once.

Rewrote some tests, and added some new tests.